included same title plugin in similiar context plugin
