+++ lib/plugins/similar-context-plugin.js
@@ -8,12 +8,15 @@ const utils = require('../utils.js')
  *
  * If param 'extractKeywords' is true: first uses the keyword-extractor to get the keywords, then compares the keywords with stringSimilarity to get a context score
  *
- * @param file - the file with its description
- * @param task - the task with its description
+ * @param sString1 - first string for comparison
+ * @param sString2 - second string for comparison
  * @param params - object with attribute extractKeywords as parameter for extracting the keywords (true - extract keyword before calculating similarity score, false - whithout extraction)
  */
 function similarityPlugin (sString1, sString2, params) {
   let extractKeywords = (params && params['extractKeywords']) || false
+
+  sString1 = utils.basename(sString1)
+  sString2 = utils.basename(sString2)
 
   if ((typeof sString1 === 'undefined') || (typeof sString2 === 'undefined')) {
     throw new Error(`Two Strings as input are needed.`)
