+++ lib/score-manager.js
@@ -133,7 +133,7 @@ function getAggregator (aggregator) {
  * task, for example the task name. It returns a floating point numeric score
  * in the range 0.0 to 1.0 which describes the degree in which the file and
  * the task are correlated in the aspect that this particular Plugin is focused
- * on. For example, the `same-title` Plugin will return 1.0 if the title of the
+ * on. For example, the `similar-context` Plugin will return 1.0 if the title of the
  * file is the same as the title of the task and 0.0 otherwise.
  *
  * An Aggregator is a policy that combines a set of scores that were previously
