+++ test/plugins/similar-context-test.js
@@ -81,6 +81,22 @@ buster.testCase('similar-context-plugin', {
     let sString2 = 'test'
     let keywordExtraction = { extractKeywords: 'test' }
     buster.assert.exception(() => plugin(sString1, sString2, keywordExtraction))
+  },
+
+  'should return 0.0 when there is no match at all ignoring the file extension': function () {
+    let file = { title: '123.jpeg' }
+    let task = { title: 'location' }
+
+    let result = plugin(file.title, task.title)
+    buster.assert.equals(result, 0.0)
+  },
+
+  'should return 1.0 when there is a total match ignoring the file extension': function () {
+    let file = { title: 'location.png' }
+    let task = { title: 'location' }
+
+    let result = plugin(file.title, task.title)
+    buster.assert.equals(result, 1.0)
   }
 
 })
