+++ test/score-manager-test.js
@@ -2,7 +2,7 @@ const buster = require('buster')
 const scoreManager = require('../lib/score-manager')
 const aggregator = require('../lib/score-aggregator')
 
-const sameTitlePlugin = require('../lib/plugins/same-title-plugin')
+const similarContextPlugin = require('../lib/plugins/similar-context-plugin')
 
 buster.testCase('ScoreManager with configuration', {
   'scoreWith': {
@@ -284,13 +284,13 @@ buster.testCase('ScoreManager with configuration', {
 })
 
 buster.testCase('ScoreManager Integration', {
-  'should be able to use sameTitlePlugin': function () {
+  'should be able to use similarContextPlugin': function () {
     let config = {
       aggregator: new aggregator.Largest(),
       plugins: {
-        'same-title': {
-          use: sameTitlePlugin,
-          inputs: ['file', 'tasks[]']
+        'similar-context': {
+          use: similarContextPlugin,
+          inputs: ['file.title', 'tasks[].title']
         }
       }
     }
@@ -300,7 +300,7 @@ buster.testCase('ScoreManager Integration', {
       file: { title: 'location.png' },
       tasks: [
         { title: 'location' },
-        { title: 'something_else' }
+        { title: '12345' }
       ]
     }
 
@@ -314,9 +314,9 @@ buster.testCase('ScoreManager Integration', {
       let config = {
         aggregator: new aggregator.Largest(),
         plugins: {
-          'same-title': {
-            use: 'same-title-plugin',
-            inputs: ['file', 'tasks[]']
+          'similar-context': {
+            use: 'similar-context-plugin',
+            inputs: ['file.title', 'tasks[].title']
           }
         }
       }
@@ -326,7 +326,7 @@ buster.testCase('ScoreManager Integration', {
         file: { title: 'location.png' },
         tasks: [
           { title: 'location' },
-          { title: 'something_else' }
+          { title: '12345' }
         ]
       }
